requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and Supabase
supabase==2.0.2
postgrest==0.13.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing and ML
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
lightgbm==4.1.0

# HTTP client for webhooks
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Optional: Plotting libraries (for chart generation)
matplotlib==3.8.2
plotly==5.17.0

# Optional: Additional utilities
python-jose[cryptography]==3.3.0  # For JWT tokens if needed
passlib[bcrypt]==1.7.4  # For password hashing if needed

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production server (alternative to uvicorn)
gunicorn==21.2.0
